Clean up iteration graph
