+++ src/tensora/iteration_graph/__init__.py
@@ -1,9 +1,3 @@
 from .definition import Definition, TensorLayer
-from .iteration_graph_to_ir import KernelType, generate_ir
-from .merge_lattice import (
-    IterationMode,
-    Lattice,
-    LatticeConjuction,
-    LatticeDisjunction,
-    LatticeLeaf,
-)
+from .generate_ir import generate_ir
+from .merge_lattice import Lattice, LatticeConjuction, LatticeDisjunction, LatticeLeaf
