+++ src/tensora/iteration_graph/identifiable_expression/__init__.py
@@ -1,5 +1,4 @@
-from .ast import Assignment, Expression, Variable  # noqa: F401
-from .exhaust_tensors import exhaust_tensor  # noqa: F401
-from .tensor_leaf import TensorLeaf  # noqa: F401
-from .to_ir import to_ir  # noqa: F401
-from .variables import to_c_code  # noqa: F401
+from .ast import Assignment, Expression, Variable
+from .exhaust_tensor import exhaust_tensor
+from .tensor_leaf import TensorLeaf
+from .to_ir import to_ir
