--- src/tensora/iteration_graph/merge_lattice/__init__.py
@@ -1,7 +1,0 @@
-from .merge_lattice import (  # noqa: F401
-    IterationMode,
-    Lattice,
-    LatticeConjuction,
-    LatticeDisjunction,
-    LatticeLeaf,
-)
